requirements.txt
pandas
pandas-gbq
numpy
requests
aiohttp
dask
gcsfs
fsspec
google-cloud-storage
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
google-cloud-bigquery
oauthlib
requests-oauthlib
gspread
facebook-business
python-dotenv